Fix compiler error message in case of of 64 bit architecture
git-svn-id: svn://svn.minisip.org/minisip/trunk@2667 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ minisip/minisip/gui/textui/MinisipTextUI.cxx
@@ -536,7 +536,7 @@ void MinisipTextUI::showTransactions(string command){
 #ifdef _MSC_VER
 							);
 #else
-							+ "  ObjecdId: " + itoa((int)*(*i)));
+							+ "  ObjecdId: " + itoa((int64_t)*(*i)));
 #endif
 					n++;
 				}
@@ -1295,7 +1295,7 @@ void MinisipTextUI::guiExecute(string cmd){
 		handled=true;
 		string arg = command.substr(2);
 		arg = trim(arg);
-		unsigned spos = arg.find(" ");
+		size_t spos = arg.find(" ");
 		if (spos==string::npos){
 			displayMessage("im usage: im <user> <message>",red);
 		}else{
